add test code for fjall notx
