+++ src/metastore/stores/mod.rs
@@ -3,3 +3,6 @@ mod fjall_notx;
 
 pub use fjall::FjallStore;
 pub use fjall_notx::FjallStoreNotx;
+
+#[cfg(test)]
+mod test_utils;
